Navigating lineages

Core functions

NCBITaxonomy.lineage — Function

lineage(tax::NCBITaxon; stop_at::NCBITaxon=ncbi"root")

Returns an array of NCBITaxon going up to the root of the taxonomy, or to the optionally specified stop_at taxonomic node.

NCBITaxonomy.children — Function

children(t::NCBITaxon)

Returns the node immediately below the taxon given as argument, or nothing if the taxon is terminal.

NCBITaxonomy.descendants — Function

descendants(t::NCBITaxon)

Recurisvely calls children until all terminal nodes under the taxon are reached. Depending on the taxonomic level, and number of taxa under the taxon considered, this can be a long function to run.

Base.parent — Function

parent(taxon::NCBITaxon)

Returns the taxon from which the argument is descended.

NCBITaxonomy.rank — Function

rank(taxon::NCBITaxon)

Returns the rank of a taxon.

NCBITaxonomy.taxonomicdistance! — Function

This function fills a pre-allocated square matrix D with the taxonomic distance between taxa in a vector tax. The keywords areguments are d (the distance as a function to closest matching level), strict (a boolean to decide whether the distance of a taxon with itself should be read from the distance dictionary, or set to 0), and other keywords passed to the lineage function.

Because the distances are symetrical, there are only (n(n-1))/2 measurements to do.

By default, this function uses the distances from Shimatani (2001):

equal rank	distance

species	0

genus	1

family	2

subclass	3

:fallback	4

Shimatani, K. 2001. On the measurement of species diversity incorporating species differences. Oikos 93:135–147.

Examples

The children function will return a list of NCBITaxon that are immediately descending from the one given as argument. For example, the genus Lamellodiscus contains:

using NCBITaxonomy

ncbi"Lamellodiscus" |> children

32-element Array{NCBITaxon,1}:
 Lamellodiscus obeliae (ncbi:119376)
 Lamellodiscus erythrini (ncbi:119377)
 Lamellodiscus virgula (ncbi:119378)
 Lamellodiscus drummondi (ncbi:119379)
 Lamellodiscus elegans (ncbi:142933)
 Lamellodiscus ignoratus (ncbi:142934)
 Lamellodiscus mormyri (ncbi:142935)
 Lamellodiscus verberis (ncbi:142936)
 Lamellodiscus baeri (ncbi:165562)
 Lamellodiscus bidens (ncbi:165563)
 ⋮
 Lamellodiscus pagrosomi (ncbi:428598)
 Lamellodiscus donatellae (ncbi:630949)
 Lamellodiscus abbreviatus (ncbi:990273)
 Lamellodiscus confusus (ncbi:990274)
 Lamellodiscus falcus (ncbi:990275)
 Lamellodiscus kechemirae (ncbi:990276)
 Lamellodiscus neifari (ncbi:990277)
 Lamellodiscus diplodi (nom. nud.) (ncbi:993027)
 Lamellodiscus theroni (ncbi:1331957)

Note that the parent function does the opposite of children:

ncbi"Lamellodiscus kechemirae" |> parent

Lamellodiscus (ncbi:119230)

To get the full descendants of a taxon (i.e. the children of its children, recursively), we can do:

descendants(ncbi"Diplectanidae")

128-element Array{NCBITaxon,1}:
 Acleotrema (ncbi:67518)
 unclassified Acleotrema (ncbi:2624465)
 Acleotrema sp. (ncbi:67519)
 Furnestinia (ncbi:92201)
 Furnestinia echeneis (ncbi:92202)
 Diplectanum (ncbi:119228)
 Diplectanum aequans (ncbi:119229)
 Diplectanum veropolynemi (ncbi:282277)
 Diplectanum penangi (ncbi:330063)
 Diplectanum umbrinum (ncbi:428596)
 ⋮
 Diplectanidae sp. AU8 (ncbi:2587167)
 Latiphagum (ncbi:2675311)
 Latiphagum setosum (ncbi:2600130)
 Dolicirroplectanum (ncbi:2691601)
 Dolicirroplectanum lacustre (ncbi:2580524)
 Aetheolabes (ncbi:2778320)
 Aetheolabes goldiensis (ncbi:2778322)
 Rhabdosynochus (ncbi:2800816)
 Rhabdosynochus viridisi (ncbi:2800817)

We can also work upwards in the taxonomy, using the lineage function – it takes an optional stop_at argument, which is the farther up it will go:

lineage(ncbi"Lamellodiscus elegans"; stop_at=ncbi"Monogenea")

6-element Array{NCBITaxon,1}:
 Monogenea (ncbi:37945)
 Monopisthocotylea (ncbi:37844)
 Dactylogyridea (ncbi:2486284)
 Diplectanidae (ncbi:67517)
 Lamellodiscus (ncbi:119230)
 Lamellodiscus elegans (ncbi:142933)

The rank function is useful to know where in the taxonomy you are:

[t => rank(t) for t in lineage(ncbi"Lamellodiscus elegans"; stop_at=ncbi"Monogenea")]

6-element Array{Pair{NCBITaxon,Symbol},1}:
              Monogenea (ncbi:37945) => :class
      Monopisthocotylea (ncbi:37844) => :subclass
       Dactylogyridea (ncbi:2486284) => :order
          Diplectanidae (ncbi:67517) => :family
         Lamellodiscus (ncbi:119230) => :genus
 Lamellodiscus elegans (ncbi:142933) => :species