Navigating lineages

Core functions

NCBITaxonomy.lineage — Function

lineage(tax::NCBITaxon; stop_at::NCBITaxon=ncbi"root")

Returns an array of NCBITaxon going up to the root of the taxonomy, or to the optionally specified stop_at taxonomic node.

NCBITaxonomy.children — Function

children(t::NCBITaxon)

Returns the node immediately below the taxon given as argument, or nothing if the taxon is terminal.

NCBITaxonomy.descendants — Function

descendants(t::NCBITaxon)

Recurisvely calls children until all terminal nodes under the taxon are reached. Depending on the taxonomic level, and number of taxa under the taxon considered, this can be a long function to run.

Examples

The children function will return a list of NCBITaxon that are immediately descending from the one given as argument. For example, the genus Lamellodiscus contains:

using NCBITaxonomy

ncbi"Lamellodiscus" |> children

32-element Array{NCBITaxon,1}:
 Lamellodiscus obeliae (119376)
 Lamellodiscus erythrini (119377)
 Lamellodiscus virgula (119378)
 Lamellodiscus drummondi (119379)
 Lamellodiscus elegans (142933)
 Lamellodiscus ignoratus (142934)
 Lamellodiscus mormyri (142935)
 Lamellodiscus verberis (142936)
 Lamellodiscus baeri (165562)
 Lamellodiscus bidens (165563)
 ⋮
 Lamellodiscus pagrosomi (428598)
 Lamellodiscus donatellae (630949)
 Lamellodiscus abbreviatus (990273)
 Lamellodiscus confusus (990274)
 Lamellodiscus falcus (990275)
 Lamellodiscus kechemirae (990276)
 Lamellodiscus neifari (990277)
 Lamellodiscus diplodi (nom. nud.) (993027)
 Lamellodiscus theroni (1331957)

To get the full descendants of a taxon (i.e. the children of its children, recursively), we can do:

descendants(ncbi"Diplectanidae")

126-element Array{NCBITaxon,1}:
 Acleotrema (67518)
 Acleotrema sp. (67519)
 Lobotrema ardens (82865)
 Furnestinia (92201)
 Furnestinia echeneis (92202)
 Diplectanum (119228)
 Diplectanum aequans (119229)
 Lamellodiscus (119230)
 Lamellodiscus obeliae (119376)
 Lamellodiscus erythrini (119377)
 ⋮
 unclassified Murraytrematoides (2636062)
 unclassified Nasobranchitrema (2637370)
 unclassified Lobotrema (2640914)
 unclassified Pseudorhabdosynochus (2645849)
 Latiphagum (2675311)
 Dolicirroplectanum (2691601)
 Lobotrema asiaticus (2730338)
 Aetheolabes (2778320)
 Aetheolabes goldiensis (2778322)

We can also work upwards in the taxonomy, using the lineage function – it takes an optional stop_at argument, which is the farther up it will go:

lineage(ncbi"Lamellodiscus elegans"; stop_at=ncbi"Monogenea")

6-element Array{NCBITaxon,1}:
 Monogenea (37945)
 Monopisthocotylea (37844)
 Dactylogyridea (2486284)
 Diplectanidae (67517)
 Lamellodiscus (119230)
 Lamellodiscus elegans (142933)

Internal functions

NCBITaxonomy._descendant_nodes — Function

_descendant_nodes(id::Vector{T}) where {T <: Int}

Internal function to retrieve the descendants of a given array of nodes.

_descendant_nodes(id::T) where {T <: Int}

Internal function to recursively get the descendants.

NCBITaxonomy._taxa_from_id — Function

_taxa_from_id(id::Vector{T}) where {T <: Int}

Get a list of NCBITaxon from a vector of ids.

NCBITaxonomy._children_nodes — Function

childrennodes(id::T) where {T <: Int}

Internal function to retrieve the id of the children of a given node.

NCBITaxonomy._parent_of — Function

_parent_of(tax::NCBITaxon)

Returns the taxon from which the argument is descended.