Navigating lineages

Core functions

NCBITaxonomy.lineage — Function

lineage(tax::NCBITaxon; stop_at::NCBITaxon=ncbi"root")

Returns an array of NCBITaxon going up to the root of the taxonomy, or to the optionally specified stop_at taxonomic node.

NCBITaxonomy.children — Function

children(t::NCBITaxon)

Returns the node immediately below the taxon given as argument, or nothing if the taxon is terminal.

NCBITaxonomy.descendants — Function

descendants(t::NCBITaxon)

Recurisvely calls children until all terminal nodes under the taxon are reached. Depending on the taxonomic level, and number of taxa under the taxon considered, this can be a long function to run.

Base.parent — Function

parent(taxon::NCBITaxon)

Returns the taxon from which the argument is descended.

NCBITaxonomy.rank — Function

rank(taxon::NCBITaxon)

Returns the rank of a taxon.

Examples

The children function will return a list of NCBITaxon that are immediately descending from the one given as argument. For example, the genus Lamellodiscus contains:

using NCBITaxonomy

ncbi"Lamellodiscus" |> children

32-element Array{NCBITaxon,1}:
 Lamellodiscus obeliae (119376)
 Lamellodiscus erythrini (119377)
 Lamellodiscus virgula (119378)
 Lamellodiscus drummondi (119379)
 Lamellodiscus elegans (142933)
 Lamellodiscus ignoratus (142934)
 Lamellodiscus mormyri (142935)
 Lamellodiscus verberis (142936)
 Lamellodiscus baeri (165562)
 Lamellodiscus bidens (165563)
 ⋮
 Lamellodiscus pagrosomi (428598)
 Lamellodiscus donatellae (630949)
 Lamellodiscus abbreviatus (990273)
 Lamellodiscus confusus (990274)
 Lamellodiscus falcus (990275)
 Lamellodiscus kechemirae (990276)
 Lamellodiscus neifari (990277)
 Lamellodiscus diplodi (nom. nud.) (993027)
 Lamellodiscus theroni (1331957)

Note that the parent function does the opposite of children:

ncbi"Lamellodiscus kechemirae" |> parent

Lamellodiscus (119230)

To get the full descendants of a taxon (i.e. the children of its children, recursively), we can do:

descendants(ncbi"Diplectanidae")

126-element Array{NCBITaxon,1}:
 Acleotrema (67518)
 unclassified Acleotrema (2624465)
 Acleotrema sp. (67519)
 Furnestinia (92201)
 Furnestinia echeneis (92202)
 Diplectanum (119228)
 Diplectanum aequans (119229)
 Diplectanum veropolynemi (282277)
 Diplectanum penangi (330063)
 Diplectanum umbrinum (428596)
 ⋮
 Diplectaninae gen. sp. 1.4 MVT-2019 (2501364)
 Diplectaninae gen. sp. 2.1 MVT-2019 (2501365)
 Diplectanidae sp. AU8 (2587167)
 Latiphagum (2675311)
 Latiphagum setosum (2600130)
 Dolicirroplectanum (2691601)
 Dolicirroplectanum lacustre (2580524)
 Aetheolabes (2778320)
 Aetheolabes goldiensis (2778322)

We can also work upwards in the taxonomy, using the lineage function – it takes an optional stop_at argument, which is the farther up it will go:

lineage(ncbi"Lamellodiscus elegans"; stop_at=ncbi"Monogenea")

6-element Array{NCBITaxon,1}:
 Monogenea (37945)
 Monopisthocotylea (37844)
 Dactylogyridea (2486284)
 Diplectanidae (67517)
 Lamellodiscus (119230)
 Lamellodiscus elegans (142933)

The rank function is useful to know where in the taxonomy you are:

[t => rank(t) for t in lineage(ncbi"Lamellodiscus elegans"; stop_at=ncbi"Monogenea")]

6-element Array{Pair{NCBITaxon,Symbol},1}:
              Monogenea (37945) => :class
      Monopisthocotylea (37844) => :subclass
       Dactylogyridea (2486284) => :order
          Diplectanidae (67517) => :family
         Lamellodiscus (119230) => :genus
 Lamellodiscus elegans (142933) => :species

Internal functions

NCBITaxonomy._descendants — Function

_descendants(id::T) where {T <: Int}

Recursively get the descendants of a given node.

NCBITaxonomy._children — Function

_children(id::T) where {T <: Int}

Internal function to retrieve the id of the children of a given node.

NCBITaxonomy._taxa_from_id — Function

_taxa_from_id(id::Vector{T}) where {T <: Int}

Get a list of NCBITaxon from a vector of ids.