Finding taxa

The taxon function

Missing docstring.

Missing docstring for taxon. Check Documenter's build log for details.

NCBITaxonomy.vernacular — Function

vernacular(t::NCBITaxon)

This function will return nothing if no vernacular name is known, and an array of names if found. It searches the "common name" and "genbank common name" category of the NCBI taxonomy name table.

NCBITaxonomy.synonyms — Function

synonyms(t::NCBITaxon)

This function will return nothing if no synonyms exist, and an array of names if they do. It returns all of the

NCBITaxonomy.authority — Function

authority(t::NCBITaxon)

This function will return nothing if no authority exist, and a string with the authority if found.

The taxon function will return a NCBITaxon object, which has two fields: name and id. We do not return the class attribute, because the package will always return the scientific name, as the examples below illustrate:

using NCBITaxonomy
taxid("Bos taurus")

There is a convenience string macro to replace the taxid function:

ncbi"Bos taurus"

Bos taurus (ncbi:9913)

Note that because the names database contains vernacular and deprecated names, the scientific name will be returned, no matter what you search

taxid("cow")

This may be a good point to note that we can use the vernacular function to get a list of NCBI-known vernacular names:

taxid("cow") |> vernacular

It also work with authorities:

taxid("cow") |> authority

You can pass an additional strict=false keyword argument to the taxon function to perform fuzzy name matching using the Levenshtein distance:

taxon("Paradiplozon homion", strict=false)

Paradiplozoon homoion (ncbi:147838)

Note that fuzzy searching comes at a performance cost, so it is preferable to use the strict matching unless necessary. As a final note, you can specify any distance function from the StringDistances package, using the dist argument.

Building a better namefilter

The taxon function, by default, searches in the entire names table. In many cases, we can restrict the scope of the search quite a lot by searching only in the range of names that match a given condition. For this reason, the taxon function also has a method with a first argument being a data frame of names. These are generated using namefilter, as well as a varitety of helper functions.

Missing docstring.

Missing docstring for namefilter. Check Documenter's build log for details.

Here is an illustration of why using namefilters makes sense. Let's say we have to search for a potentially misspelled name:

@time taxon("Ebulavurus"; strict=false);

Ebolavirus (ncbi:186536)

We can use the virusfilter() function to generate a table with viruses only:

viruses = virusfilter()
@time taxon(viruses, "Bumbulu ebolavirus"; strict=false);

Bombali ebolavirus (ncbi:2010960)

A namefilter can be built in a number of ways, including by passing a list of taxa:

diplectanids = namefilter(descendants(ncbi"Diplectanidae"))
taxon(diplectanids, "Lamellodiscus")

Lamellodiscus (ncbi:119230)

Standard namefilters

To save some time, there are namefilters pre-populated with the large-level taxonomic divisions.

NCBITaxonomy.bacteriafilter — Function

bacteriafilter()

Returns a namefinder limited to the bacterial division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

NCBITaxonomy.virusfilter — Function

virusfilter()

Returns a namefinder limited to the viral division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments. Note that phage are covered by phagefinder.

NCBITaxonomy.mammalfilter — Function

mammalfilter(;inclusive::Bool=true)

Returns a namefinder limited to the mammal division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

If the keyword argument inclusive is set to false, this will not search for organisms assigned to a lower division, in this case rodents (covered by rodentfinder) and primates (covered by primatefinder). The default behavior is to include these groups.

NCBITaxonomy.vertebratefilter — Function

vertebratefilter(;inclusive::Bool=true)

Returns a namefinder limited to the vertebrate division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

If the keyword argument inclusive is set to false, this will not search for organisms assigned to a lower division, in this case mammals (covered by mammalfinder). The default behavior is to include these groups, which also include the groups covered by mammalfinder itself.

NCBITaxonomy.plantfilter — Function

plantfilter()

Returns a namefinder limited to the "plant and fungi" division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

NCBITaxonomy.invertebratefilter — Function

invertebratefilter()

Returns a namefinder limited to the invertebrate division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

Note that this is limited organisms not covered by plantfinder, bacteriafinder, and virusfinder.

NCBITaxonomy.rodentfilter — Function

rodentfilter()

Returns a namefinder limited to the rodent division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

NCBITaxonomy.primatefilter — Function

primatefilter()

Returns a namefinder limited to the primate division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

NCBITaxonomy.phagefilter — Function

phagefilter()

Returns a namefinder limited to the phage division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

NCBITaxonomy.environmentalsamplesfilter — Function

environmentalsamplesfilter()

Returns a namefinder limited to the environmental samples division of the NCBI taxonomy. See the documentation for namefinder and taxid for more information about arguments.

All of these return a dataframe which can be passed to the taxon function as a first argument.