Finding taxa

The taxid function

NCBITaxonomy.taxid — Function

taxid(name::T; fuzzy::Bool = false) where {T <: String}

Returns the taxonomic ID of a taxon, given as a string. This function searches in the entire names table, which is unlikely to give a good performance when using fuzzy matching. We encourage the use of the namefinder function to build a custom version.

The taxid function will return a NCBITaxon object, which has two fields: name and id. We do not return the class attribute, because the package will always return the scientific name, as the examples below illustrate:

using NCBITaxonomy
taxid("Bos taurus")

Bos taurus (9913)

There is a convenience string macro to replace the taxid function:

ncbi"Bos taurus"

Bos taurus (9913)

Note that because the names database contains vernacular and deprecated names, the scientific name will be returned, no matter what you search

taxid("cow")

Bos taurus (9913)

You can pass an additional fuzzy=true keyword argument to the taxid function to perform fuzzy name matching using the Levenshtein distance:

taxid("Paradiplozon homion", fuzzy=true)

Paradiplozoon homoion (147838)

Note that both fuzzy searching and non-standard naming come at a performance cost, so it is preferable to use the strict matching unless necessary.

Building a better namefinder

The namefinder function has one job: generating a function that works exactly like taxid, but only searches on a smaller subset of the data. In fact, taxid is a special case of namefinder, which simply searches the whole database.

NCBITaxonomy.namefinder — Function

namefinder(df::T) where {T <: DataFrame}

Generates a name-finding function that takes a string as an argument, and a boolean (fuzzy) as a keyword argument to switch from strict to fuzzy matching. The generated function will return a NCBITaxon for a given string. By default, the function taxid is working on the entire names table, which is going to be slow if there are many names to fuzzy match. The keyword argument verbose (default to false) indicates whether the distance to a fuzzy match must be printed.

Under strict matching, if no match is found, the namefinder will return nothing. This can be used to switch to the fuzzy namefinder automatically. The fuzzy namefinder uses the findnearest function from StringDistances, which will always return something. The compare function from the same package may be used to see how similar the names are, and to decide whether to keep them.

Altough the input dataframe is supposed to be a subset of the (unexported) names_table, all that is required is that it has the columns tax_id, name, and class. Make of that information what you wish...

NCBITaxonomy.descendantsfinder — Function

descendantsfinder(t::NCBITaxon)

Returns a namefinder for all taxa below the one given as argument. This function calls descendants internally, so it might not be the optimal way when dealing with large groups.

Here is an illustration of why using namefinders makes sense. Let's say we have to search for a potentially misspelled name:

@time taxid("Evolavirus"; fuzzy=true)

Ebolavirus (186536)

We can build a more efficient namefinder by selecting the nodes in the taxonomy that belong to the VRL division. Doing so requires to call namefinder on a DataFrame. Note that we are doing some merging here, which results in the data frame we use having more columns than the names data frame – but this does not matter, because the namefinder is not picky about having too much information.

using DataFrames, DataFramesMeta
viralfinder = namefinder(
         join(
           @where(
               select(NCBITaxonomy.nodes_table, [:tax_id, :division_code]),
               :division_code .== Symbol("VRL")
           ),
           NCBITaxonomy.names_table;
           on = :tax_id
         )
       )

@time viralfinder("Evolavirus"; fuzzy=true);

Ebolavirus (186536)

For searches in specific groups, the descendantsfinder is a convenient wrapper: it will return a namefinder limited to all taxa below its argument.

diplectanidfinder = descendantsfinder(taxid("Diplectanidae"))
diplectanidfinder("Lamellodiscus")

Lamellodiscus (119230)

Internal functions

NCBITaxonomy._get_sciname_from_taxid — Function

_get_sciname_from_taxid(df::T, id::Int) where {T <: DataFrame}

This internal function will return a scientific name from a numerical id.

NCBITaxonomy._df_from_taxlist — Function

_df_from_taxlist(tax::Vector{NCBITaxon})

Returns a subset of the names dataframe based on a vector of taxa.